Bug 1918490 - Move support link element outside of label element. r=reusable-components-reviewers,accessibility-frontend-reviewers,desktop-theme-reviewers,hjones,nlapre,sidebar-reviewers,nsharpley
The motivation for this change is to improve the screen reader user
experience when navigating through moz-checkbox and moz-radio elements.
When a screen reader encounters and focuses a moz-checkbox or
moz-radio element, only the label is announced. Previously the label
and the support-link element would be announced which is confusing since
it is not clear what will happen when activating one of these moz
elements.

Additionally, we remove the display flex and width rules from
sidebar-customize.css to prevent sidebar test breakage and create a
larger click surface for moz-radio and moz-checkbox when used in the
sidebar.

The moz-input-common.css change makes moz-label behave more like XUL
labels where the click area for the label and control span the entire
width unless there is a support link in the moz-label. In this case,
clicking on the other side of the support link will not activate the
respective control.

Differential Revision: https://phabricator.services.mozilla.com/D226181

UltraBlame original commit: 25a8e51c3fe2b657ba449fba790f9d2ef57da784