Bug 1932513 - Fix crashing redesigned toolbar related UI tests r=ohorvath,aaronmt
The were 4 UI tests which use the redesigned toolbar that started to fail all of a sudden after [[ https://bugzilla.mozilla.org/show_bug.cgi?id=1908330 | this ]] change landed.

The problem was caused by the fact that the UI test were launched with the redesigned toolbar being **disabled** and after the tests were actually launched the feature was disabled using the applySettingsExceptions function.
Because of this, the layout of the toolbar was in a "weird" state, crashing the app.

To overcome this problem, I've updated the composeTestRule in both HomeScreenWithNavbarTest and TabbedBrowsingWithNavbarTest to enable the feature right from the beginning.
I've also had to create a diffrent function to open the tab tray and use it in tabsCounterShortcutMenuOptionFromNavbarInPrivateModeTest and verifyPrivateTabsTrayWithOpenTabFromNavbarTest UI tests.

All UI tests from TabbedBrowsingWithNavbarTest and HomeScreenWithNavbarTest successfully passed 3x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D229762

UltraBlame original commit: 54a7b37d7f39b99201c11dd9b5f6ac895aa3816f