Bug 1917000 - Turn "userScripts" into optional-only permission r=zombie
Mark "userScripts" as an optional-only permission and update all tests
that expect it to be a required permission, which is not supported.

This also improves the implementation to account for the fact that a
permission can be granted post install.

The disabling of functionality upon revoking the permission will be in
the next patch.

Differential Revision: https://phabricator.services.mozilla.com/D229712

UltraBlame original commit: 9a6e72ab04f915be5363f15f375759befc11a9e1