Bug 1932800 - Add a check to SelfIsSelectable to preserve editor drag and drop behavior. r=masayuki
We were relying on editable text frames not being empty. We fixed that
on the previous patches but this caused editor d&d to break.

For now this preserves the selection behavior. In the future we should
look into removing the IsEmpty() condition or so, probably, or make it
more subtle.

Differential Revision: https://phabricator.services.mozilla.com/D229996

UltraBlame original commit: ee6527afae3bd18f674dc662e6c4677a00d7ce96