Bug 1930281: Use external macroassembler in BaselineCodeGen r=jandem
StackMacroAssembler can only be stack-allocated on the main thread, so we need to be able to use a different MacroAssembler that can enforce GC safety off-thread (similarly to IonHeapMacroAssembler).

I wanted to put a StackMacroAssembler as a member of BaselineInterpreterGenerator, to avoid having to allocate one at the call site. However, it was uninitialized in the BaselineCodeGen construtor, because derived class fields are initialized after calling the superclass constructor.

Differential Revision: https://phabricator.services.mozilla.com/D229850

UltraBlame original commit: 88a4914c152bc719e501ba80f6efcdcceed76454