Bug 1933623 - testRecentlyClosedTabGroupsSingleTab consistency r=dao
This patch matches the tab title listed on the menu item in the "recently closed tabs" menu with the tab title of the restored tab.

This test was arbitrarily restoring the first listed tab in a menu, but the expectation was that it would load a specific tab. This failed nondeterministically. This patch should always work regardless of how many or which tabs are contained in the tab group.

Differential Revision: https://phabricator.services.mozilla.com/D230346

UltraBlame original commit: 4c8e6bcf7d23ae06521e81548e05c75d5253bd84