Bug 1931799. If the display list builder (and hence it's arena) is doomed we don't need to bother tracking freeing individual items to the arena. r=gfx-reviewers,nical$layout-reviewers
If the arena is going to go away we don't need to free each allocation, all that does is append it to a free list, which takes time, and then we'll have to destroy that free list too. The arena will de-allocate the memory the same.

This mirrors what we do for frames which also use the same arena (nsPresArena):
https://searchfox.org/mozilla-central/rev/f8488adfb9c3e25a80ba6a21e04950d58165e956/layout/base/PresShell.h#291

This does mean that we don't write the nsPresArena poison over the memory before we release it back to the OS (we are left with whatever zero-ing or poisoning is done or not done by the allocator at a lower level then nsPresArena).

Differential Revision: https://phabricator.services.mozilla.com/D229291

UltraBlame original commit: 3e38c8e1415d6fc93cee6b64473baf16cffbfe03