Bug 1934238 - Don't try to position anything from XUL if the widget is not in the normal sizemode. r=rkraesig
When the window is already properly maximized (which can happen with the
pre-skeleton UI), we shouldn't adjust the position of the window.

The reason this worked before my patch was because the skeleton UI
didn't have the right bounds, so the position didn't change.

But I don't think we should move the window at all in this case (I
believe the current patch in bug 1866583 part 4 maybe fix this too, but
it's better to be explicit).

Differential Revision: https://phabricator.services.mozilla.com/D230694

UltraBlame original commit: 4b396c6161bbd8f6f0c4b26ef5608bea9272ec0e