Bug 1918267 - [devtools] Avoid waiting for replies of hanging processes in DevTools communications made to content processes. r=devtools-reviewers,jdescottes
DevTools currently broadcast its data to all processes (in order to reach all service workers, which may run in arbitrary processes).
We expect for the queries to be processed in the content process before replying back to the client/frontend in the related RDP request.

But, in some case, some content process may be stuck on a CC or slow page script.
When this happens, DevTools would be blank on opening or very slow to operate.

Let's try to avoid waiting for any query send to process which timed out on any past query.

Differential Revision: https://phabricator.services.mozilla.com/D229960

UltraBlame original commit: e7660a1ea32adf14c8a4d7d5c6c05e59fbf4cb54