Bug 1913757 - Try to reuse LifoAllocs for Ion compilations. r=iain
When an Ion compilation task is finished, we add it to the runtime's `ionFreeTaskBatch_`.
With this patch we check this batch and try to reuse LifoAllocs.

This improves some Speedometer 3 subtests by a few percent on Android.

Differential Revision: https://phabricator.services.mozilla.com/D230750

UltraBlame original commit: 8e478167d68b2ce4c8c9d42d89ef16f94c46a8da