Bug 1934972 part 3 - Store used JitZone stubs in WarpSnapshot, remove zoneStubsToReadBarrier_. r=iain,jonco
`zoneStubsToReadBarrier_` was used to delay read-barriers for the `JitZone` stubs
for off-thread Ion compilations. This only works well if we cancel Ion compilations
on GC.

This patch changes this to instead record used stubs in the Warp snapshot that we trace
for off-thread compilations.

We also no longer generate the string-concat stub eagerly but only if the CacheIR has
a string-concat op.

Differential Revision: https://phabricator.services.mozilla.com/D231222

UltraBlame original commit: af214c46fc64b03d5ab6d078417f10da66f59a16