Bug 1935547 part 3: Add "stretch" to our mochitest invalid_values lists for various length-valued properties that don't accept css-sizing keywords. r=layout-reviewers,emilio
This is essentially a late-breaking followup for earlier changeset
https://hg.mozilla.org/mozilla-central/rev/7496f354f4f2 which added this
keyword as an allowed value in some properties, but neglected to add it as a
disallowed value.

Per https://drafts.csswg.org/css-sizing-4/#sizing-values , the 'stretch'
keyword is only valid for width, height, and their min/max properties. It's
also valid for properties that are defined to take the same values as those
properties -- e.g. these properties' logical aliases, and 'flex-basis' and
'flex' which accept <'width'> as part of their syntax.

The 'stretch' keyword is not valid for other <length>-valued properties, in
general, so it's worth asserting that we don't accept it there; hence this
patch which adds it to a smattering of properties.

Differential Revision: https://phabricator.services.mozilla.com/D231340

UltraBlame original commit: a0443b9f1b4217364bd8f209eb74bbfd703b002f