Bug 1933423. Only create nsDisplayFixedPosition if painting to the window. r=botond
The only thing that useFixedPosition controls is whether we create a nsDisplayFixedPosition item. The only things that nsDisplayFixedPosition does different are things that only affect painting to the window (CreateWebRenderCommands, UpdateScrollData, ShouldGetFixedAnimationId).

I don't think this causes any behaviour difference, but we only need a nsDisplayFixedPosition item if we are painting to the window.

Differential Revision: https://phabricator.services.mozilla.com/D230243

UltraBlame original commit: c1135593470ae05016ad0109f8310783f94b9096