Bug 1936061 - Always evaluate the Nimbus default to know if microsurveys should be enabled r=android-reviewers,twhite
Previously the app would start with an uninitialized Nimbus value for the navigation bar
experiment defaulting to `false` and keep this in memory for the entire lifecycle.
Now we will always evaluate the Nimbus default if a value is not persisted on disk.

Differential Revision: https://phabricator.services.mozilla.com/D231575

UltraBlame original commit: 0f3367748be21be450b7dc935f486246bf5ed2d3