Bug 1933053 - QM: Add a stub nsIQuotaManagerServiceInternal interface and integrate it into QuotaManagerService; r=dom-storage-reviewers,asuth
This patch introduces the nsIQuotaManagerServiceInternal interface as a stub
for future internal functionality of the QuotaManagerService. The interface is
currently empty but serves as a foundation for an additional method that will
be added in a subsequent patch.

Differential Revision: https://phabricator.services.mozilla.com/D231484

UltraBlame original commit: 4b42bdd5104f06ae6f36dad01a2eb023a5ce2f2d