Bug 1933053 - QM: Cache thumbnail private identity id; r=dom-storage-reviewers,asuth
The nsIQuotaUtilsService interface is restricted to the main thread, so the
thumbnail private identity id is retrieved on the main thread in the handler
for the sessionstore-windows-restored observer topic. It is not retrieved in
profile-do-change or profile-after-change handlers because those notifications
occur very early during startup, while the ContextualIdentityService can be
initialized after user events start being handled (this is done for performance
reasons and the start of user events handling is signaled by firing the
sessionstore-windows-restored notification).

Additionally, retrieving the id earlier could trigger the initialization of
many JS modules in background tasks, which we want to avoid.

Furthermore, some Necko-related tests expect that the cookies service remains
uninitialized after calling do_get_profile. If the id were retrieved during
profile-do-change, it would inadvertently initialize the cookies service, as
the ContextualIdentityService needs to traverse cookies while creating default
identities. This would conflict with test expectations.

So, for some time after startup, the thumbnail private identity id may not be
known. This is acceptable, as we do not expect specific origin operations to be
triggered early. Even if they are, it would only mean that some optimizations,
like counting thumbnail private identity temporary origins, are not utilized.

Note: xpcshell tests currently do not fire sessionstore-windows-restored during
do_get_profile, so the thumbnail private identity id will not be known in those
tests. However, this functionality can be added later if needed.

Differential Revision: https://phabricator.services.mozilla.com/D231265

UltraBlame original commit: 530c54805289f923b75e248c82f0a35b5c4fbe87