Bug 1913111 - wasm: Store inlined caller offset information on TrapSite. r=jseward
Add a sparse map of inlineCallerOffsets to TrapSites, similar to
CallSites. This map lives on the inner TrapSitesForKind class as
it is indexed by an entry in the vector.

A later patch will use this information and fill it in.

Depends on D231765

Differential Revision: https://phabricator.services.mozilla.com/D231766

UltraBlame original commit: 5a3a452002e3288f214c77b6f4a7c533b68263ab