Bug 1935511 - Rename getMaxVersionRecords r=translations-reviewers,gregtatum
Renames the getMaxVersionRecords() function to
getMaxSupportedVersionRecords() because I think
it improves overall clarity and readability.

The record that is retrieved may not be the absolute
maximum record, but only the maximum that is supported
in this version of the source code.

Depends on D232081

Differential Revision: https://phabricator.services.mozilla.com/D231316

UltraBlame original commit: 8d7743d3408ae0ae10dbe13c45184ae97a773e14