Bug 1936163 - Rework isLangTagSupported for Translations r=translations-reviewers,gregtatum
This patch reworks the isLangTagSupported functionality on the
TranslationsParent to instead find the normalized-for-Translations
language tag that matches the provided language tag argument.

The prior use cases that required boolean functionality are preserved
by the new functions returning `null` if a compatible langTag is not found.

Now that we will be supporting languages that may be differentiated by a
script tag, it is imperative that we have a unified method of retrieving
the exact language tag expected by the Translations model.

Depends on D231632

Differential Revision: https://phabricator.services.mozilla.com/D231884

UltraBlame original commit: 3533bf1d6cf84ffeaae6d8ae19612fb71a53d00c