Bug 1937624 part 3: Add variants of WPT for 'stretch' size keyword on abspos element, with the abspos element having a static position. r=dshin
This forces the browser to compute the static position, which is a bit
complicated in this situation (since the static position comes from doing
layout in the in-flow parent, whose writing mode is potentially different from
the abspos child itself and also different from the abspos containing block).

In Gecko at least, this leads to a special case where we're forced to resolve
the abspos element's block-size a bit earlier-than-we-would-otherwise, as part
of the writing-mode conversion, here:
https://searchfox.org/mozilla-central/rev/66aaf96f955d9bdc4bdb087a5075d5fbff2b9c8d/layout/generic/ReflowInput.cpp#1508-1538

This patch's "stretch-block-size-003.html" testcase provides code-coverage for
that codepath, to ensure we're handling the 'stretch' keyword properly
there. And the inline-size variant is just included for symmetry/completeness.

I'm generating these testcases as modified variants of their -002 siblings,
where I've removed the explicit `0` insets, but I've also removed the padding
on their box-tree parent.  The intent is to make their static position exactly
the same as the absolute position in the -002 test, so that the expected sizing
(when stretching to reach the abspos container's far edge) will still be the
same.

Differential Revision: https://phabricator.services.mozilla.com/D232270

UltraBlame original commit: 4c5fe40750673492e6ecc0cc9ce2778bbe7a5572