Bug 1934489 - P1: Update ScrollPosition cache when "scrollable" element has no scrollable frame. r=morgan
A document's body element can lose its scroll frame if the root element (eg. <html>) is restyled to overflow scroll/auto. In that case we will not get any useful notifications for the body element except for a reframe to a non-scrolling frame.

The place where we see this as a problem is in hit testing because
accessibles get erroneously clipped out because of stale scroll
positions.

Differential Revision: https://phabricator.services.mozilla.com/D231821

UltraBlame original commit: b1c9cdccd0dd5f124702b98ad8e93c3068b57801