Bug 1916849 - Implement intrinsic sizing for spanning grid items with flex tracks. r=tlouw,dholbert
This has a couple unfortunate parts to its implementation. In particular, the
adjustment to ShouldApplyAutoMinSize are difficult to explain other than that
function doesn't actually calculate if the minimum contribution of an item is
content-based, but rather checks if the caller should also call MinSize or not.

Part of this calculation includes checking the min-size-auto of the item, but
some are just matching checks for special cases in MinSize.

This could be optimized, and should definitely be refactored to simplify how
all of this is computed.

This still doesn't quite correctly handle all cases for items spanning flex
tracks and tracks with a definite length, but that case is already not handled
correctly so this won't introduce any regressions for that case unless the
expected result is that such an item will have an intrinsic size of zero. In
that case, simply not calculating anything would accidentally be the correct
value, and this new calculation may or may not produce that result.

Differential Revision: https://phabricator.services.mozilla.com/D230829

UltraBlame original commit: 07405ecf81266689f32276234769b841df89d8c1