Bug 1931445 - Track base code address separately for backtracking; r=iain
As a first step to only storing the 4-byte offsets on the backtrack
stack this patch splits out the base code address and adds it to the
backtrack offset popped from the stack. The base code address needs to
be patched in during code generation.

Differential Revision: https://phabricator.services.mozilla.com/D232665

UltraBlame original commit: 0890a7e83d999845db86a36d46887bf86275e793