Bug 1937814 [wpt PR 49721] - [scope] Expand :scope selectors for :has() cache, a=testonly
Automatic update from web-platform-tests
[scope] Expand :scope selectors for :has() cache

The :scope pseudo-class can refer to different elements depending on
context. This is similar to the '&' pseudo-class, but differs
in that :scope refers to a single *element*, not a specific selector list.

This causes problems for the :has() cache, since it uses
the serialization of a selector as the key. In CL:5680906, this was
solved by expanding '&' in the cache-key, and this CL now does a similar
trick for :scope. Unlike '&', however, the :scope selector itself does
not know what it refers to; it must come from context. We can use
the pointer value of the scope node for this, since the :has() cache
only exists on the stack during style recalc.

Fixed: 383343312
Change-Id: I08e5d3caf18831d614caed474081d32caa43bd64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6090905
Reviewed-by: Byungwoo Lee <bleeigalia.com>
Commit-Queue: Anders Hartvoll Ruud <andruudchromium.org>
Cr-Commit-Position: refs/heads/main{#1397188}

--

wpt-commits: b9ff3b3521376a66d4765bd2765928c549d3df09
wpt-pr: 49721

UltraBlame original commit: 2eafb0a3957a68cb8eba7fa09a75ff7d74257f59