Bug 1939620 Part 2 - Make test_drag_caret_from_front_to_end_across_columns more stable. r=layout-reviewers,webdriver-reviewers,emilio
This patch reverts the edit in Bug 1932800 - Fix caret baseline of empty inline
and text frames.
https://hg.mozilla.org/mozilla-central/rev/b0a2ef9d7ae3b2d9f421f339b420dd54cb0df44b

Additionally, convert `<p>` to `<div>` to remove the margin at the beginning of
the first column and the margin at the end of the second column. This ensures
that the first and second columns look the same.

In a contenteditable div, an image can be selected by dragging a cursor mode
AccessibleCaret across it. To avoid triggering AccessibleCaret's selection mode
during cursor mode dragging tests, we make the image non-selectable.

Differential Revision: https://phabricator.services.mozilla.com/D233067

UltraBlame original commit: 41f89b32973f3691344bca0a8811d1039939ac2d