Bug 1919554: Test DND inside/between shadow DOMs r=edgar
These patches include a new DND test that performs DND from/to elements in a
shadow DOM and it found that we weren't properly propagating these events
outside a shadow DOM when the drag source and target were in the same frame.
They were instead only sent to the target element itself.  We therefore end up
sending a confusing stream of events to the elements -- for example, the
test found that the document received dragover events without first receiving
a dragenter.  And no drop or dragleave may ever arrive after a dragover.

The reason is that we set the relatedTarget for dragenter, dragexit and
dragleave events (Bug 1420590), which the spec doesn't call for.  The world
seems to want it, and has since convinced some other browsers to do the same
thing.  So people like it and undoing it would probably break real things.

We don't have a fix for the failure, which is filed as bug 1932955, so
this test is currently blocked.  A fix for that bug could end up invalidating
the assumptions of this test but it is currently useful as a way to reproduce
the issue.

Differential Revision: https://phabricator.services.mozilla.com/D229983

UltraBlame original commit: ae5c743353a9e42650a2684538a24c1c7a2bf8ff