Bug 1923701 - Declare threads.h as a system header. r=firefox-build-system-reviewers,sergesanspaille
The new version of wasm2c uses it under some conditions. Of course, this
conflicts with Threads.h on case-insensitive file systems, but it seems
that one is leftover from ancient times and not used anymore, so we can
remove it.

Differential Revision: https://phabricator.services.mozilla.com/D233306

UltraBlame original commit: 3572def9b19f39df96fa38d6335234bec041f231