Bug 1936294 - Fix-up mFocusedWindow / mFocusedElement when replacing frameloaders. r=dom-core,sefeng
The issue is that when replacing frame loaders for a non-remote browser,
mFocusedElement is not that <browser> element. It instead points to the
(potentially null) focused element in mFocusedWindow.

If mFocusedWindow is about to be destroyed (or put into the bfcache), clear it
and point mFocusedElement to the <browser> instead, so that we can focus the
new frame loader if needed later.

If we don't do this, the focused window gets cleared when the window is nuked
in nsFocusManager::WasNuked (but by that point it's too late to do any fix up,
as we've lost the context that we're swapping, and of our frameloader owner,
and everything else).

Test in D232437.

Differential Revision: https://phabricator.services.mozilla.com/D233444

UltraBlame original commit: d3f21fe8379a32487f9cb76cff247707e44feafe