Bug 1930378: Ensure nsIClearDataService.cleanupAfterDeletionAtShutdown is always called after sanitization.r=dom-storage-reviewers,janv,asuth,emz
We ensure that nsIClearDataService.cleanupAfterDeletionAtShutdown always get called at the end of sanitization process.
For optimization reasons, this was only performed conditionally where it used to check if sanitization was performed or
not and skips cleanup if it was not. Now, the condition is more general and relaxed where we check the existence of 'to-be-removed'
directory and check its' contents to decide if cleanup is required. It was harder to ensure that cleanup was done correctly for all
cases with prior approach.

Differential Revision: https://phabricator.services.mozilla.com/D231268

UltraBlame original commit: 0c98ae0803c80777d1b33a214cb712324d1178eb