Bug 1938317 part 3 - Use same optimization in BacktrackingAllocator::installAllocationsInLIR. r=jseward
Very similar to the optimization in `populateSafepoints`, but for the non-call safepoints list.
This avoids quadratic behavior in `findFirstNonCallSafepoint`.

When compiling the module in the bug report we spent a lot of time under `installAllocationsInLIR`
and with this patch almost all of that is eliminated.

Differential Revision: https://phabricator.services.mozilla.com/D233355

UltraBlame original commit: a1e2fec0520a08209ebd636ff0562a9d45177b9a