Bug 1934689 - make moz-page-nav-button icons optional r=reusable-components-reviewers,tgiles
This patch makes it so that moz-page-nav-button works with or without an icon. It also updates the documentation to reflect this change, and adds a test to cover the nav's responsive display, focusing on the differences between navs that use icons and those that dont.

Differential Revision: https://phabricator.services.mozilla.com/D231814

UltraBlame original commit: 9f753492d47eab1c2e7755279f1373625dfea6cf