Bug 1940501 [wpt PR 49971] - CanvasRenderingContext2DSettings: follow the spec types, a=testonly
Automatic update from web-platform-tests
CanvasRenderingContext2DSettings: follow the spec types

The type for willReadFrequently is boolean, not an enum with values of
"true", "false", and "undefined".

Add WPT tests for this. And for the rest of the attributes.

The WPT tests failed for the desynchronized property because, when
it was discovered that this feature was never implemented on macOS,
it was disabled at compile time, when parsing the input from V8,
losing the information required to correctly implement
getContextAttributes. Add a temporary workaround for this situation
by separately storing the specified desynchronized value.

Bug: 388290815
Change-Id: I00e6b6c82a23cdb627a3eba1314ebf3b0f9abf9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6157723
Reviewed-by: Yi Xu <yiyixchromium.org>
Commit-Queue: ccameron chromium <ccameronchromium.org>
Reviewed-by: Aaron Krajeski <aaronhkchromium.org>
Cr-Commit-Position: refs/heads/main{#1404356}

--

wpt-commits: d431fd42b36dfa5a67f08d0927bbb1f65e40eca4
wpt-pr: 49971

UltraBlame original commit: 14ba0826f62cff96e20e952e10da19eb52023dc2