Bug 1933428 - Part 3: Recompute partitionKey for top-level documents if required. r=bvandersloot,smaug
The principal of a document might be changed during loading in certain
situation, e.g. a sandboxed document. In this case, we need to recompute
the partitionKey of the document to reflect the change on the principal.

Differential Revision: https://phabricator.services.mozilla.com/D230819

UltraBlame original commit: 32e0f0c78f145e92ec039ff81abd7773cd1abc48