Bug 1922335 - Implement basic support for DataSourceSurface-based snapshotting of old contents. r=nical
This implements some basic snapshotting functionality, and hooks it to
the ::view-transition-old pseudo-elements.

Some tests that now fail because we start rendering the old snapshot,
but not at the right position, as for now the view transition pseudo
tree is all statically-positioned. That's all expected for now until we
implement the right styling for them, the top layer behavior, etc.

Main remaining question is how to get a hand on the right layer manager
from nsImageFrame::Destroy.

Differential Revision: https://phabricator.services.mozilla.com/D231977

UltraBlame original commit: d8203347358e65effea26444b2f8f62d815118e6