Bug 1941731 - Bail out on the parent side if ForwardError finds a DeviceLost error. r=webgpu-reviewers,ErichDonGubler
A common pattern in WebGPUParent.cpp is to early return if ForwardError returns true which happens when it enconters an error other than DeviceLost. Teo suggested that DeviceLost should also cause the early-return which makes sense as far as I can tell so this patch implements the suggestion. As a result, a DeviceLost error while atempting a device will prevent us from continuing to run code that assumes the device has been successfully inserted in the registry.

In addition, some unnecessary code in RcvAdapterRequestDevice was removed. LoseDevice isn't needed since we haven't created the device in the first place, and the assertion was redundant with the code in ForwardError.

Differential Revision: https://phabricator.services.mozilla.com/D234473

UltraBlame original commit: 6b28ddc86237acaa9957934a3b105f88f9de6c26