Bug 1901492 - Simplify some require-trusted-types-for checks. r=smaug
Currently, there is only one sink group defined in the spec, accepted
by nsCSPParser::handleRequireTrustedTypesForDirective and used by
callers of GetTrustedTypesCompliantString. This means that the flag
added in bug 1909168 to indicate whether a policy contains a
require-trusted-types-for directive equivalently indicates whether
a policy contains a require-trusted-types-for for the supported
sink group and we currently don't need to browse the list of
directives to perform actual sink group matching.

Differential Revision: https://phabricator.services.mozilla.com/D233505

UltraBlame original commit: 23ca1773d73a13c012b1cc635fe6615da46a4675