Bug 1942735 - Update the swipe gesture Modifier when the underlying state object has been updated. r=android-reviewers,npoon
As part of `Modifier.pointerInput`, we are able to provide a `key`, similar to LaunchedEffect, to tell the Modifier to restart whenever that `key` updates. By changing this from `Unit` to our `state` object, the gestures should now be properly enabled/disabled based on our gesture business logic (i.e. disabled when there is vertical scrolling or the Tabs Tray is in multi-selection).

Differential Revision: https://phabricator.services.mozilla.com/D235213

UltraBlame original commit: fed3206664a81d1830242d0172f4e1919210692b