Bug 1942349 - Add support for model variants; r=translations-reviewers,settings-reviewers,mossop,nordzilla
This is a big patch, but it adds the support for model variants by
defining a language pair struct instead of a fromLanguage and
toLanguage. This struct can also include a source variant and target
variant. The code will resolve the pivot models as needed.

This is not going to be release user facing, but can be used in a
Nightly setting. There is nothing enforcing this, but it's the intention
of the code.

This will allow us to test multiple variants of models in Nightly and
gather feedback from users and other engineers.

Differential Revision: https://phabricator.services.mozilla.com/D235495

UltraBlame original commit: fe82f7593d69f7752d9b8fe2dd91d7285fb9082b